re-add link.json
Since v0.2.0 the `link.json` schema is no longer part of JSON-FG. The file should still be published for now on beta.schemas.opengis.net as an informative schema in case any other schema includes a reference. All references to the `link.json` schema should be changed as the schema is deprecated.